Integrate the new unit testing framework
