+++ libcaf_io/caf/io/network/receive_buffer.hpp
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <algorithm>
 #include <cstddef>
 #include <cstring>
 #include <limits>
