+++ libcaf_io/test/io-test.cpp
@@ -1,13 +1,8 @@
-#define CAF_TEST_NO_MAIN
+// This file is part of CAF, the C++ Actor Framework. See the file LICENSE in
+// the main distribution directory for license terms and copyright or visit
+// https://github.com/actor-framework/actor-framework/blob/master/LICENSE.
 
-#include "caf/test/unit_test_impl.hpp"
+#include "caf/io/middleman.hpp"
+#include "caf/test/caf_test_main.hpp"
 
-#include "io-test.hpp"
-
-int main(int argc, char** argv) {
-  using namespace caf;
-  init_global_meta_objects<id_block::io_test>();
-  io::middleman::init_global_meta_objects();
-  core::init_global_meta_objects();
-  return test::main(argc, argv);
-}
+CAF_TEST_MAIN(caf::io::middleman)
