+++ libcaf_net/test/detail/rfc6455.cpp
@@ -2,11 +2,9 @@
 // the main distribution directory for license terms and copyright or visit
 // https://github.com/actor-framework/actor-framework/blob/master/LICENSE.
 
-#define CAF_SUITE detail.rfc6455
-
 #include "caf/detail/rfc6455.hpp"
 
-#include "net-test.hpp"
+#include "caf/test/test.hpp"
 
 #include "caf/byte_buffer.hpp"
 #include "caf/span.hpp"
@@ -18,172 +16,170 @@
 
 using namespace caf;
 
-namespace {
+using impl = detail::rfc6455;
 
-struct fixture {
-  using impl = detail::rfc6455;
+SUITE("detail.rfc6455") {
 
-  auto bytes(std::initializer_list<uint8_t> xs) {
-    byte_buffer result;
-    for (auto x : xs)
-      result.emplace_back(static_cast<std::byte>(x));
-    return result;
-  }
-
-  template <class T>
-  auto take(const T& xs, size_t num_bytes) {
-    auto n = std::min(xs.size(), num_bytes);
-    return std::vector<typename T::value_type>{xs.begin(), xs.begin() + n};
-  }
-};
-
-} // namespace
-
-BEGIN_FIXTURE_SCOPE(fixture)
-
-TEST_CASE("masking") {
-  auto key = uint32_t{0xDEADC0DE};
-  auto data = bytes({0x12, 0x34, 0x45, 0x67, 0x89, 0x9A});
-  auto masked_data = data;
-  MESSAGE("masking XORs the repeated key to data");
-  impl::mask_data(key, masked_data);
-  CHECK_EQ(masked_data, bytes({
-                          0x12 ^ 0xDE,
-                          0x34 ^ 0xAD,
-                          0x45 ^ 0xC0,
-                          0x67 ^ 0xDE,
-                          0x89 ^ 0xDE,
-                          0x9A ^ 0xAD,
-                        }));
-  MESSAGE("masking masked data again gives the original data");
-  impl::mask_data(key, masked_data);
-  CHECK_EQ(masked_data, data);
+auto bytes(std::initializer_list<uint8_t> xs) {
+  byte_buffer result;
+  for (auto x : xs)
+    result.emplace_back(static_cast<std::byte>(x));
+  return result;
 }
 
-TEST_CASE("decoding a frame with RSV bits fails") {
+template <class T>
+auto take(const T& xs, size_t num_bytes) {
+  auto n = std::min(xs.size(), num_bytes);
+  return std::vector<typename T::value_type>{xs.begin(), xs.begin() + n};
+}
+
+TEST("masking") {
+  auto key = uint32_t{0xDEADC0DE};
+  auto data = bytes({0x12, 0x34, 0x45, 0x67, 0x89, 0x9A});
+  SECTION("masking XORs the repeated key to data") {
+    auto masked_data = data;
+    impl::mask_data(key, masked_data);
+    check_eq(masked_data, bytes({
+                            0x12 ^ 0xDE,
+                            0x34 ^ 0xAD,
+                            0x45 ^ 0xC0,
+                            0x67 ^ 0xDE,
+                            0x89 ^ 0xDE,
+                            0x9A ^ 0xAD,
+                          }));
+  }
+  SECTION("masking masked data again gives the original data") {
+    auto masked_data = data;
+    impl::mask_data(key, masked_data);
+    impl::mask_data(key, masked_data);
+    check_eq(masked_data, data);
+  }
+}
+
+TEST("decoding a frame with RSV bits fails") {
   std::vector<uint8_t> data;
   byte_buffer out = bytes({
     0xF2, // FIN + RSV + binary frame opcode
     0x00, // data size = 0
   });
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), -1);
+  check_eq(impl::decode_header(out, hdr), -1);
 }
 
-TEST_CASE("no mask key and no data") {
+TEST("decode a header with no mask key and no data") {
   std::vector<uint8_t> data;
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0, as_bytes(make_span(data)), out);
-  CHECK_EQ(out, bytes({
+  check_eq(out, bytes({
                   0x82, // FIN + binary frame opcode
                   0x00, // data size = 0
                 }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 2);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0u);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 2);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0u);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("valid mask key and no data") {
+TEST("decode a header with valid mask key but no data") {
   std::vector<uint8_t> data;
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0xDEADC0DE,
                        as_bytes(make_span(data)), out);
-  CHECK_EQ(out, bytes({
+  check_eq(out, bytes({
                   0x82,                   // FIN + binary frame opcode
                   0x80,                   // MASKED + data size = 0
                   0xDE, 0xAD, 0xC0, 0xDE, // mask key,
                 }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 6);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0xDEADC0DE);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 6);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0xDEADC0DE);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("no mask key plus small data") {
+TEST("decode a header with no mask key plus small data") {
   std::vector<uint8_t> data{0x12, 0x34, 0x45, 0x67};
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0, as_bytes(make_span(data)), out);
-  CHECK_EQ(out, bytes({
+  check_eq(out, bytes({
                   0x82,                   // FIN + binary frame opcode
                   0x04,                   // data size = 4
                   0x12, 0x34, 0x45, 0x67, // masked data
                 }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 2);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0u);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 2);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0u);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("valid mask key plus small data") {
+TEST("decode a header with valid mask key plus small data") {
   std::vector<uint8_t> data{0x12, 0x34, 0x45, 0x67};
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0xDEADC0DE,
                        as_bytes(make_span(data)), out);
-  CHECK_EQ(out, bytes({
+  check_eq(out, bytes({
                   0x82,                   // FIN + binary frame opcode
                   0x84,                   // MASKED + data size = 4
                   0xDE, 0xAD, 0xC0, 0xDE, // mask key,
                   0x12, 0x34, 0x45, 0x67, // masked data
                 }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 6);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0xDEADC0DE);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 6);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0xDEADC0DE);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("no mask key plus upper bound on small data") {
+TEST("decode a header with no mask key plus upper bound on small data") {
   std::vector<uint8_t> data;
   data.resize(125, 0xFF);
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0, as_bytes(make_span(data)), out);
-  CHECK_EQ(take(out, 6), bytes({
+  check_eq(take(out, 6), bytes({
                            0x82,                   // FIN + binary frame opcode
                            0x7D,                   // data size = 125
                            0xFF, 0xFF, 0xFF, 0xFF, // masked data
                          }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 2);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0u);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 2);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0u);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("valid mask key plus upper bound on small data") {
+TEST("decode a header with valid mask key plus upper bound on small data") {
   std::vector<uint8_t> data;
   data.resize(125, 0xFF);
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0xDEADC0DE,
                        as_bytes(make_span(data)), out);
-  CHECK_EQ(take(out, 10), bytes({
+  check_eq(take(out, 10), bytes({
                             0x82,                   // FIN + binary frame opcode
                             0xFD,                   // MASKED + data size = 125
                             0xDE, 0xAD, 0xC0, 0xDE, // mask key,
                             0xFF, 0xFF, 0xFF, 0xFF, // masked data
                           }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 6);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0xDEADC0DE);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 6);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0xDEADC0DE);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("no mask key plus medium data") {
+TEST("decode a header with no mask key plus medium data") {
   std::vector<uint8_t> data;
   data.resize(126, 0xFF);
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0, as_bytes(make_span(data)), out);
-  CHECK_EQ(take(out, 8),
+  check_eq(take(out, 8),
            bytes({
              0x82,                   // FIN + binary frame opcode
              0x7E,                   // 126 -> uint16 size
@@ -191,20 +187,20 @@ TEST_CASE("no mask key plus medium data") {
              0xFF, 0xFF, 0xFF, 0xFF, // first 4 masked bytes
            }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 4);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0u);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 4);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0u);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("valid mask key plus medium data") {
+TEST("decode a header with valid mask key plus medium data") {
   std::vector<uint8_t> data;
   data.resize(126, 0xFF);
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0xDEADC0DE,
                        as_bytes(make_span(data)), out);
-  CHECK_EQ(take(out, 12),
+  check_eq(take(out, 12),
            bytes({
              0x82,                   // FIN + binary frame opcode
              0xFE,                   // MASKED + 126 -> uint16 size
@@ -213,19 +209,19 @@ TEST_CASE("valid mask key plus medium data") {
              0xFF, 0xFF, 0xFF, 0xFF, // first 4 masked bytes
            }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 8);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0xDEADC0DE);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 8);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0xDEADC0DE);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("no mask key plus upper bound on medium data") {
+TEST("decode a header with no mask key plus upper bound on medium data") {
   std::vector<uint8_t> data;
   data.resize(65535, 0xFF);
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0, as_bytes(make_span(data)), out);
-  CHECK_EQ(take(out, 8),
+  check_eq(take(out, 8),
            bytes({
              0x82,                   // FIN + binary frame opcode
              0x7E,                   // 126 -> uint16 size
@@ -233,20 +229,20 @@ TEST_CASE("no mask key plus upper bound on medium data") {
              0xFF, 0xFF, 0xFF, 0xFF, // first 4 masked bytes
            }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 4);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0u);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 4);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0u);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("valid mask key plus upper bound on medium data") {
+TEST("decode a header with valid mask key plus upper bound on medium data") {
   std::vector<uint8_t> data;
   data.resize(65535, 0xFF);
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0xDEADC0DE,
                        as_bytes(make_span(data)), out);
-  CHECK_EQ(take(out, 12),
+  check_eq(take(out, 12),
            bytes({
              0x82,                   // FIN + binary frame opcode
              0xFE,                   // 126 -> uint16 size
@@ -255,19 +251,19 @@ TEST_CASE("valid mask key plus upper bound on medium data") {
              0xFF, 0xFF, 0xFF, 0xFF, // first 4 masked bytes
            }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 8);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0xDEADC0DE);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 8);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0xDEADC0DE);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("no mask key plus large data") {
+TEST("decode a header with no mask key plus large data") {
   std::vector<uint8_t> data;
   data.resize(65536, 0xFF);
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0, as_bytes(make_span(data)), out);
-  CHECK_EQ(take(out, 14),
+  check_eq(take(out, 14),
            bytes({
              0x82, // FIN + binary frame opcode
              0x7F, // 127 -> uint64 size
@@ -275,20 +271,20 @@ TEST_CASE("no mask key plus large data") {
              0xFF, 0xFF, 0xFF, 0xFF, // first 4 masked bytes
            }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 10);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0u);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 10);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0u);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-TEST_CASE("valid mask key plus large data") {
+TEST("decode a header with valid mask key plus large data") {
   std::vector<uint8_t> data;
   data.resize(65536, 0xFF);
   byte_buffer out;
   impl::assemble_frame(impl::binary_frame, 0xDEADC0DE,
                        as_bytes(make_span(data)), out);
-  CHECK_EQ(take(out, 18),
+  check_eq(take(out, 18),
            bytes({
              0x82, // FIN + binary frame opcode
              0xFF, // MASKED + 127 -> uint64 size
@@ -297,11 +293,11 @@ TEST_CASE("valid mask key plus large data") {
              0xFF, 0xFF, 0xFF, 0xFF, // first 4 masked bytes
            }));
   impl::header hdr;
-  CHECK_EQ(impl::decode_header(out, hdr), 14);
-  CHECK_EQ(hdr.fin, true);
-  CHECK_EQ(hdr.mask_key, 0xDEADC0DE);
-  CHECK_EQ(hdr.opcode, impl::binary_frame);
-  CHECK_EQ(hdr.payload_len, data.size());
+  check_eq(impl::decode_header(out, hdr), 14);
+  check_eq(hdr.fin, true);
+  check_eq(hdr.mask_key, 0xDEADC0DE);
+  check_eq(hdr.opcode, impl::binary_frame);
+  check_eq(hdr.payload_len, data.size());
 }
 
-END_FIXTURE_SCOPE()
+} // SUITE("detail.rfc6455")
