+++ libcaf_test/caf/test/then.hpp
@@ -4,13 +4,14 @@
 
 #pragma once
 
+#include "caf/detail/test_export.hpp"
 #include "caf/test/block.hpp"
 #include "caf/test/block_type.hpp"
 
 namespace caf::test {
 
 /// Represents a `THEN` block.
-class then : public block {
+class CAF_TEST_EXPORT then : public block {
 public:
   using block::block;
